scripts: add scripts that show package and recipe
This script shows the packages that are used and where the recipe /
bb-file for the particular build exists. To use this one has to do a
first build with:
	bitbake -g <image>
so that the "task-depends.dot" file is generated. If you run the script
from the folder where task-depends exist, then you don't have to
provide a file argument. If you run the script from another folder, then
you will have to provide the full path to the task-depends.dot file.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>