recipes: add a page containing information about recipes
In some cases it can be useful to have a page describing certain things
related to layers and recipes. This servers as a starting point for that
and there is also a helper script provided that helps users to figure
out what packages actually are being built and which of all copies of
bb-files for that package that actually is used.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>